Nettoyages dans Blowfish

+++ trunk/blowfish.c
@@ -65,7 +65,7 @@ Modifications by Xavier.Leroy@inria.fr, 2005. (Marked "XL".)
 
 #define N               16
 
-static const unsigned long ORIG_P[16 + 2] = {
+static const u32 ORIG_P[16 + 2] = {
         0x243F6A88L, 0x85A308D3L, 0x13198A2EL, 0x03707344L,
         0xA4093822L, 0x299F31D0L, 0x082EFA98L, 0xEC4E6C89L,
         0x452821E6L, 0x38D01377L, 0xBE5466CFL, 0x34E90C6CL,
@@ -73,7 +73,7 @@ static const unsigned long ORIG_P[16 + 2] = {
         0x9216D5D9L, 0x8979FB1BL
 };
 
-static const unsigned long ORIG_S[4][256] = {
+static const u32 ORIG_S[4][256] = {
     {   0xD1310BA6L, 0x98DFB5ACL, 0x2FFD72DBL, 0xD01ADFB7L,
         0xB8E1AFEDL, 0x6A267E96L, 0xBA7C9045L, 0xF12C7F99L,
         0x24A19947L, 0xB3916CF7L, 0x0801F2E2L, 0x858EFC16L,
@@ -336,9 +336,9 @@ static const unsigned long ORIG_S[4][256] = {
 /* XL: turned F into a macro + endianness handling */
 
 #if 0
-static unsigned long F(BLOWFISH_CTX *ctx, unsigned long x) {
+static u32 F(BLOWFISH_CTX *ctx, u32 x) {
    unsigned short a, b, c, d;
-   unsigned long  y;
+   u32  y;
 
    d = (unsigned short)(x & 0xFF);
    x >>= 8;
@@ -363,12 +363,12 @@ static unsigned long F(BLOWFISH_CTX *ctx, unsigned long x) {
 
 #endif
 
-void Blowfish_Encrypt(BLOWFISH_CTX *ctx, unsigned long *xl, unsigned long *xr){
-  unsigned long  Xl;
-  unsigned long  Xr;
+void Blowfish_Encrypt(BLOWFISH_CTX *ctx, u32 *xl, u32 *xr){
+  u32  Xl;
+  u32  Xr;
 
 #if 0
-  unsigned long  temp;
+  u32  temp;
   short       i;
 
   Xl = *xl;
@@ -415,12 +415,12 @@ void Blowfish_Encrypt(BLOWFISH_CTX *ctx, unsigned long *xl, unsigned long *xr){
 #endif
 }
 
-void Blowfish_Decrypt(BLOWFISH_CTX *ctx, unsigned long *xl, unsigned long *xr){
-  unsigned long  Xl;
-  unsigned long  Xr;
+void Blowfish_Decrypt(BLOWFISH_CTX *ctx, u32 *xl, u32 *xr){
+  u32  Xl;
+  u32  Xr;
 
 #if 0
-  unsigned long  temp;
+  u32  temp;
   short       i;
 
   Xl = *xl;
@@ -472,7 +472,7 @@ void Blowfish_Decrypt(BLOWFISH_CTX *ctx, unsigned long *xl, unsigned long *xr){
 
 void Blowfish_Init(BLOWFISH_CTX *ctx, unsigned char *key, int keyLen) {
   int i, j, k;
-  unsigned long data, datal, datar;
+  u32 data, datal, datar;
 
   for (i = 0; i < 4; i++) {
     for (j = 0; j < 256; j++)
